pyproject.toml
[tool.poetry]
name = "fetch-rewards-ch"
version = "0.1.0"
description = ""
authors = ["Connor Heraty <connor.heraty.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
snowflake-connector-python = "^3.7.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
